Refactor: Move UploadManager.uploadUserActivities queries into ActivitiesRepository
