+++ plan.md
@@ -1,0 +1,20 @@
+1. **Create ActivityData DTO**
+   - Create a new data class `ActivityData` in `org.ole.planet.myplanet.repository.ActivitiesRepository` interface (or a separate file if needed, but it's small so inside the file is fine) that contains `activityId`, `userId`, and `serialized` (JsonObject).
+
+2. **Update ActivitiesRepository Interface**
+   - Add a suspend function `getPendingLoginActivities(context: Context): List<ActivityData>` to the `ActivitiesRepository` interface.
+   - Add a suspend function `markActivityUploaded(activityId: String?, responseJson: JsonObject)` to the `ActivitiesRepository` interface.
+
+3. **Update ActivitiesRepositoryImpl**
+   - Implement `getPendingLoginActivities(context: Context)` which will use `databaseService.withRealm` to find pending login activities (`isNull("_rev")` and `equalTo("type", "login")`), filter out guests, and map them to `ActivityData`.
+   - Implement `markActivityUploaded(activityId: String?, responseJson: JsonObject)` which will use `databaseService.executeTransactionAsync` to update the activity's `_id` and `_rev` using `changeRev` (or similar logic).
+
+4. **Update UploadManager.uploadUserActivities**
+   - Modify `uploadUserActivities` to call `activitiesRepository.getPendingLoginActivities(context)` instead of directly querying Realm.
+   - Modify the loop inside `uploadUserActivities` to call `activitiesRepository.markActivityUploaded(activityData.activityId, responseBody)` instead of the direct Realm transaction async block.
+
+5. **Run Pre-commit Checks**
+   - Call `pre_commit_instructions` and follow them to ensure tests pass and everything is good.
+
+6. **Submit Code**
+   - Commit and submit.
